lns_web/backend/requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.1
django-environ==0.4.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
PyJWT==1.7.1
pytz==2018.9